+++ NEWS.md
@@ -5,6 +5,11 @@
 * Added support for fixed coordinates (i.e., the aspect component of `coord_equal()`, `coord_fixed()`, `coord_map()`, `coord_quickmap()`, `coord_sf()`).
 * Added support for `geom_sf()` and `coord_sf()`.
 
+## CHANGES
+
+* Upgraded to plotly.js v1.27.0 -- https://github.com/plotly/plotly.js/releases/tag/v1.27.0
+* The `traces` argument in the `style()` function now defaults to `NULL` (instead of 1). Meaning that, by default, supplied attributes now modify _every_ trace (instead of the first one).
+
 ## Bug fixes 
 
 * When creating private plots via `api_create()`, both the plot and the data behind the plot are private (fixes #976).