+++ NEWS.md
@@ -12,7 +12,8 @@
 
 ## Bug fixes 
 
-* When creating private plots via `api_create()`, both the plot and the data behind the plot are private (fixes #976).
+* When creating private plots (via `api_create()`), both the plot and the data behind the plot are private (fixes #976).
+* Creating a plot with multiple traces (or frames) via (via `api_create()`) no longer creates multiple grids (fixes #1004).
 * Fixed bug when highlight multiple 'simple key' traces (fixes #974) 
 
 # 4.6.0