+++ NEWS.md
@@ -1,10 +1,11 @@
-# 4.6.0.9000
+# 4.7.0
 
 ## NEW FEATURES & IMPROVEMENTS
 
 * Added support for fixed coordinates (i.e., the aspect component of `coord_equal()`, `coord_fixed()`, `coord_map()`, `coord_quickmap()`, `coord_sf()`).
 * Added support for `geom_sf()` and `coord_sf()`.
-* The `api_create()` function gains a new `fileopt` argument inspired from the `fileopt` argument in `plotly_POST()` (fixes #976).
+* The `api_create()` function gains a new `fileopt` argument, which is inspired from the `fileopt` argument in the (deprecated) `plotly_POST()` function (fixes #976). It currently supports to values: `"new"` and `"overwrite"`. The default, `"overwrite"`, will overwrite existing file(s) with a matching `filename`.
+* The `filename` argument in `api_create()` now accepts a character vector of length 2, the first string is used to name the plot, and the second is used to name the grid (i.e., data behind the plot).
 
 ## CHANGES
 
@@ -13,9 +14,10 @@
 
 ## Bug fixes 
 
+* The typed array polyfill is now included *before* the plotly.js bundle, which should fix some rendering issues in some browsers, including RStudio (fixes #1010).
 * When creating private plots (via `api_create()`), both the plot and the data behind the plot are private (fixes #976).
 * Creating a plot with multiple traces (or frames) via (via `api_create()`) no longer creates multiple grids (fixes #1004).
-* Fixed bug when highlight multiple 'simple key' traces (fixes #974) 
+* Fixed bug when highlight multiple 'simple key' traces (fixes #974).
 
 # 4.6.0